Establish database connection as a side-effect
I keep going back and forth whether this module should just export
a single function which would establish the database connection
on the demand if it's not been established yet. The problem is
that you'd have to carry the db creds in every module you want to
use the connection in. That's why I decided to connect in a
separate 'init' method, and provide the db instance directly
as an export. There's a caveat: Now the connection should be
established in the project entrypoint file, once.